% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.r
\name{add_articles}
\alias{add_articles}
\title{Copy articles from one articleset to another}
\usage{
add_articles(articleset, articles, conn = conn_from_env())
}
\arguments{
\item{articleset}{The id of the articleset to which the articles will be added}

\item{articles}{A numeric vector with article ids}

\item{conn}{The connection object from \code{\link{amcat_connect}}.}
}
\value{
the articleset id
}
\description{
Add the given article ids to an existing article set
}
