% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{add_articles_to_set}
\alias{add_articles_to_set}
\title{Add articles to an article set}
\usage{
add_articles_to_set(project, articles, conn = conn_from_env(),
  articleset = NULL, articleset.name = NULL, articleset.provenance = NULL)
}
\arguments{
\item{project}{the project to add the articles to}

\item{articles}{a vector of article ids}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{articleset}{the article set id of an existing set}

\item{articleset.name}{the name for a new article set}

\item{articleset.provenance}{a provenance text for a new article set}
}
\value{
The articleset id of the new or existing article set
}
\description{
Add the given article ids to a new or existing article set
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl')

h = get_hits(queries=c("tyrant OR brute", "saddam"), labels=c("tyrant", "saddam"), sets=10271)
articles = h$id[h$query == "tyrant"]

## make new setwith the (two) documents that mention "tyrant"
setid = add_articles_to_set(project=429, articles=articles, articleset.name="New set from howto")

## set contains only these two documents
get_articles(project = 429, articleset=setid)
}
}
