% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.r
\name{add_users}
\alias{add_users}
\title{Link articlesets to the project}
\usage{
add_users(project, articles, conn = conn_from_env())
}
\arguments{
\item{project}{The id of the project to which the articlesets will be linked}

\item{articles}{A numeric vector with article ids}

\item{conn}{The connection object from \code{\link{amcat_connect}}.}
}
\value{
the articleset id
}
\description{
Link articlesets to the project
}
