% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.r
\name{amcat.add.parents}
\alias{amcat.add.parents}
\title{Add parents to a data frame of codes}
\usage{
amcat.add.parents(conn, codes, fields, columns = NULL, maxdepth = 1)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{codes}{a data frame of codes, e.g. from \code{\link{amcat.codingjob.results}}}

\item{fields}{the coding field specifications, e.g. from \code{\link{amcat.codingjob.results}}}

\item{columns}{a character vector containing the column names to add parents to}

\item{maxdepth}{how many parent columns to add for each column, e.g. maxdepth=1 gives roots only}
}
\description{
New columns are generated for each code, containing their roots / parents
If columns is not specified, adds parents to each codebook column
}
