% Generated by roxygen2 (4.0.2): do not edit by hand
\name{amcat.cast.sparse.matrix}
\alias{amcat.cast.sparse.matrix}
\title{Cast data.frame to sparse matrix}
\usage{
amcat.cast.sparse.matrix(rows, columns, values = NULL)
}
\arguments{
\item{rows}{a vector of row indices: [i,]}

\item{columns}{a vector of column indices: [,j]}

\item{values}{a vector of the values for each (non-zero) cell: [i,j] = value}
}
\value{
a sparse matrix of the dgTMatrix class (\code{\link{Matrix}} package)
}
\description{
Create a sparse matrix from matching vectors of row indices, column indices and values
}

