% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.r
\name{amcat.codingjob.results}
\alias{amcat.codingjob.results}
\title{Get the results of a codingjob}
\usage{
amcat.codingjob.results(conn, project, job)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{project}{project ID}

\item{job}{coding job ID}
}
\description{
(note: current implementation is fairly slow because it makes HTTP calls per coded article)
(it is presumably still quicker than the human coders, though :) )
}
