% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{amcat.connect}
\alias{amcat.connect}
\title{Connect to the AmCAT API}
\usage{
amcat.connect(host, token = NULL, disable_ipv6 = TRUE,
  ssl.verifypeer = FALSE, passwordfile = "~/.amcatauth")
}
\arguments{
\item{host}{the hostname, e.g. http://amcat.vu.nl or http://localhost:8000}

\item{token}{an existing token to authenticate with. If given, username and password are not used and the token is not tested}

\item{passwordfile}{optionally, specify a different password file}

\item{disable_ipv6.}{If True, only use ipv4 resolving (faster if ipv6 causes timeout). Defaults to true, but this may change in the future.}
}
\value{
A list with authentication information that is used by the other functions in this package
}
\description{
Connect to the AmCAT API and requests a temporary (24h) authentication token that will be stored in the output
The host should be known in the ~/.amcatauth file, you can use amcat.save.password to add a password to this file
}
