% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amcatr.r
\name{amcat.flush}
\alias{amcat.flush}
\title{Ask AmCAT to flush the elasticsearch}
\usage{
amcat.flush(conn)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}
}
\description{
Ask AmCAT to flush the elasticsearch
}

