% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.r
\name{amcat.get.codingjob.results}
\alias{amcat.get.codingjob.results}
\title{Get the results of a codingjob}
\usage{
amcat.get.codingjob.results(conn, project, job)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{project}{project ID}

\item{job}{coding job ID}
}
\description{
Get the results of a codingjob
}
