% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amcat.getpages}
\alias{amcat.getpages}
\title{Get and rbind pages from the AmCAT API}
\usage{
amcat.getpages(conn, path, format = "csv", page = 1, page_size = 1000,
  filters = NULL, post = FALSE, max_page = NULL)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{path}{the path of the url to retrieve (using the host from conn)}

\item{filters:}{a named vector of filters, e.g. c(project=2, articleset=3)}

\item{post}{use HTTP POST instead of GET}

\item{page:}{the page number to start retrieving}

\item{page_size:}{the number of rows per page}

\item{max_page:}{the page number to stop retrieving at, if given}
}
\value{
dataframe
}
\description{
Get and rbind pages from the AmCAT API
}

