% Generated by roxygen2 (4.0.2): do not edit by hand
\name{amcat.gettokens}
\alias{amcat.gettokens}
\title{Get Tokens from AmCAT}
\usage{
amcat.gettokens(conn, project = NULL, articleset = NULL,
  module = "corenlp_lemmatize", keep = NULL, drop = c("id", "sentence",
  "offset"), count = TRUE, filters = NULL, page_size = 1, page = 1,
  npages = NULL, sentence = NULL)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{project}{id of the project containing the tokens}

\item{articleset}{id of the articleset to get features from. If not specified, specify sentence for 'ad hoc' parsing}

\item{module}{the NLP preprocessing module to get the tokens from}

\item{keep}{an optional list of attributes to keep (and aggregate on), e.g. c("lemma", "pos1")}

\item{drop}{an optional list of attributes to drop, default (token)id and sentence. If keep is given, drop is ignored.}

\item{count}{if true, tokens will be made unique and a 'freq' column will be added}

\item{filters}{Additional filters, ie c(pos1="V", pos1="A") to select only verbs and adjectives}

\item{page_size}{the number of features (articles?) to include per call}

\item{sentence}{a sentence (string) to be parsed if articleset id is not given}
}
\value{
A data frame of tokens
}
\description{
Get Tokens (pos, lemma etc) from AmCAT
}

