% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amcatr.r
\name{amcat.getURL}
\alias{amcat.getURL}
\title{Retrieve a single URL from AmCAT with authentication and specified filters (GET or POST)}
\usage{
amcat.getURL(conn, path, filters = NULL, post = FALSE,
  post_options = list(), error_unless_200 = TRUE, null_on_404 = FALSE,
  binary = FALSE)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{path}{the path of the url to retrieve (using the host from conn)}

\item{filters}{a named vector of filters, e.g. c(project=2, articleset=3)}

\item{post}{use HTTP POST instead of GET}
}
\value{
the raw result
}
\description{
Retrieve a single URL from AmCAT with authentication and specified filters (GET or POST)
}

