% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amcat.hierarchy.cats}
\alias{amcat.hierarchy.cats}
\title{Add categories (ancestors) to a codebook 'hierarchy'}
\usage{
amcat.hierarchy.cats(hierarchy, maxdepth = 2)

amcat.hierarchy.cats(hierarchy, maxdepth = 2)
}
\arguments{
\item{hierarchy}{the hierarchy data frame from \code{\link{amcat.gethierarchy}}}

\item{maxdepth}{the maxium number of ancestors per code}

\item{hierarchy}{the hierarchy data frame from \code{\link{amcat.gethierarchy}}}

\item{maxdepth}{the maxium number of ancestors per code}
}
\value{
The hierarchy data frame with a column added for each code

The hierarchy data frame with a column added for each code
}
\description{
Adds one or more categories to codebook codes. Suppose that you have a hierarchy like

Adds one or more categories to codebook codes. Suppose that you have a hierarchy like
}
\details{
code, parent
issue, NA
economy, issue
unemployment, economy
environment, issue

The first category or 'root' for all objects will be 'issue'. The second category would be
'economy' for economy and unemployment, and 'environment' for environment. For 'issue', the second
category would simply be issue:

code, parent, cat1, cat2
issue, NA, issue, issue
economy, issue, issue, economy
unemployment, economy, issue, economy
environment, issue, issue, environment
#'

code, parent
issue, NA
economy, issue
unemployment, economy
environment, issue

The first category or 'root' for all objects will be 'issue'. The second category would be
'economy' for economy and unemployment, and 'environment' for environment. For 'issue', the second
category would simply be issue:

code, parent, cat1, cat2
issue, NA, issue, issue
economy, issue, issue, economy
unemployment, economy, issue, economy
environment, issue, issue, environment
#'
}

