% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.r
\name{amcat.hits}
\alias{amcat.hits}
\title{Conduct a query on amcat}
\usage{
amcat.hits(conn, queries, labels = queries, sets, minimal = T,
  warn.no.result = T, ...)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{queries}{a vector of queries to run}

\item{labels}{if given, labels corresponding to the queries. Alternatively, if a query starts with a label and a hashtag (e.g., label# "term1 term2"), this label is used.}

\item{sets}{one or more article set ids to query on}

\item{...}{additional arguments to pass to the AmCAT API, e.g. extra filters}
}
\value{
A data frame with hits per article
}
\description{
This function is similar to using the 'show article list' function in AmCAT. It allows you to specify a
number of queries and get document metadata and number of hits per document
}

