% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpipe.r
\name{amcat.nlpipe}
\alias{amcat.nlpipe}
\title{Process an articleset with nlpipe}
\usage{
amcat.nlpipe(conn, project, articleset, module,
  nlpipe_server = getOption("nlpiper.server", default =
  "http://localhost:5001"))
}
\arguments{
\item{conn}{the amcatr connection object}

\item{project}{the amcat project id}

\item{articleset}{the amcat articleset id}

\item{module}{the nlpipe module name}

\item{nlpipe_server}{the nlpipe server name or folder}
}
\description{
This requires vanatteveldt/nlpiper to be install as well
}

