% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amcat.runaction}
\alias{amcat.runaction}
\title{Run an action from the AmCAT API}
\usage{
amcat.runaction(conn, action, format = "csv", ...)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{action}{the name of the action}

\item{format}{the format to request, e.g. csv or json}

\item{...}{any additional (e.g. action-specific) arguments to pass to the action API}
}
\value{
A dataframe containing the result of the action [WvA: shouldn't that depend on the action??]
}
\description{
Run an action from the AmCAT API
}

