% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/amcatr.r
\name{amcat.save.password}
\alias{amcat.save.password}
\title{Add or change an entry in the cached authentication file}
\usage{
amcat.save.password(host, username, password)
}
\arguments{
\item{host}{the host, e.g. "https://amcat.nl"}

\item{username}{the AmCAT username to use}

\item{password}{the password for the AmCAT user}
}
\description{
amcat-r uses the file '~/.amcatauth' to read credentials for connecting to servers
to prevent passwords from appearing in script files. This function will update
the .amcatauth file (if present) to add or change the password for the given host
}

