% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.r
\name{amcat.strip.codeid}
\alias{amcat.strip.codeid}
\title{Strip the codeid from columns}
\usage{
amcat.strip.codeid(codes, columns = NULL)
}
\arguments{
\item{codes}{A vector containing codes, or a data frame with codes, e.g. from  \code{\link{amcat.codingjob.results}}}

\item{columns}{If codes is a data frame: the columns to strip. If not given, strips all factor columns in the data frame.}
}
\description{
Removes the ID from columns containing "<ID> - <Label>" values
}
