% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{amcat.tokens.unique_indices}
\alias{amcat.tokens.unique_indices}
\title{Create unique indices on the}
\usage{
amcat.tokens.unique_indices(tokens, context.column = "aid",
  index.columns = intersect(colnames(tokens), c("clause_id", "source_id",
  "coref")))
}
\arguments{
\item{tokens}{a data frame with tokens, which should have an aid column}

\item{context.column}{the column name with the document (context) identifier, default aid}

\item{index.columns}{a vector of column names to be made globally unique}
}
\value{
the tokens data frame with the columns made unique
}
\description{
Create unique indices on the
}

