% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amcat.upload.articles}
\alias{amcat.upload.articles}
\title{Upload new articles to AmCAT}
\usage{
amcat.upload.articles(conn, project, articleset, headline, text, date, medium,
  ...)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{project}{the project to add the articles to}

\item{articleset}{the article set id of an existing set. You can use \code{\link{amcat.add.articles.to.set}}
to create a new articleset (with articles=NULL)}

\item{headline}{the headlines of the articles to upload}

\item{text}{the text of the articles to upload}

\item{medium}{the medium of the articles to upload. Can be either of length 1 or of same length as headlines}

\item{...}{and additional fields to upload, e.g. author, byline etc.}
}
\description{
Upload articles into a given project and article set
}

