% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{amcat_connect}
\alias{amcat_connect}
\title{Connect to the AmCAT API}
\usage{
amcat_connect(host, token = NULL, disable_ipv6 = TRUE,
  ssl.verifypeer = FALSE, passwordfile = "~/.amcatauth")
}
\arguments{
\item{host}{the hostname, e.g. http://amcat.vu.nl or http://localhost:8000}

\item{token}{an existing token to authenticate with. If given, username and password are not used and the token is not tested}

\item{disable_ipv6}{If True, only use ipv4 resolving (faster if ipv6 causes timeout). Defaults to true, but this may change in the future.}

\item{ssl.verifypeer}{If True, verifies the authenticity of the peer's certificate}

\item{passwordfile}{optionally, specify a different password file}
}
\value{
A list with authentication information that is used by the other functions in this package
}
\description{
Connect to the AmCAT API and requests a temporary (24h) authentication token that will be stored in the output
The host should be known in the ~/.amcatauth file, you can use save_amcat_password to add a password to this file
}
\examples{
\dontrun{
host = 'https://amcat.nl'  ## existing and available host
username = 'XXX'           ## registered username  
password = 'XXX'           ## registered password

## first, store username and password for a given host. 
save_amcat_password(host = host, username = username, password = password)

## connect by just giving the host
conn = amcat_connect(host)
}
}
