\name{amcatr-package}
\alias{amcatr-package}
\alias{amcatr}
\docType{package}
\title{
Client bindings for the AmCAT API
}
\description{
Client bindings for the AmCAT API. More info?
}
\details{
\tabular{ll}{
Package: \tab amcatr\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-01-17\cr
License: \tab MIT\cr
}
}
\author{
Wouter van Atteveldt and Kasper Welbers

Maintainer: Wouter van Atteveldt <wouter@vanatteveldt.com>

}
\references{
Wouter van Atteveldt (2008), Semantic Network Analysis: Techniques for Extracting, Representing, and Querying Media Content (dissertation). Charleston, SC: BookSurge
}

\keyword{ package }
\seealso{
}
\examples{

}
