% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/django.r
\name{CharField}
\alias{CharField}
\title{Create a Django Form Field}
\usage{
CharField(
  label = NA,
  initial = "",
  required = T,
  min_length = NA,
  max_length = NA,
  help_text = NA
)
}
\description{
Use to create field objects for R plugins in AmCAT. See djangoFormFields().
}
