% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/django.r
\name{ChoiceField}
\alias{ChoiceField}
\title{Create a Django Form Field}
\usage{
ChoiceField(label = NA, choices, required = T, help_text = NA)
}
\description{
Use to create field objects for R plugins in AmCAT. See djangoFormFields().
}

