% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{conn_from_env}
\alias{conn_from_env}
\title{Obtain AmCAT connection from environment}
\usage{
conn_from_env()
}
\description{
! This only works if \link{amcat_connect} has been used in the current session.
}
\examples{
\dontrun{
conn = conn_from_env()
}
}
