% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{conn_to_env}
\alias{conn_to_env}
\title{Save AmCAT connection as environment variable}
\usage{
conn_to_env(conn)
}
\arguments{
\item{conn}{An AmCAT connection, as created with \link{amcat_connect}}
}
\description{
Save AmCAT connection as environment variable
}
\examples{
\dontrun{
conn = amcat_connect('localhost:8000', 'username')
conn_to_env(conn)
}
}
