% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{create_articles}
\alias{create_articles}
\title{Upload new articles to AmCAT}
\usage{
create_articles(data, project, articleset = NULL, new_set = NULL,
  provenance = NULL, conn = conn_from_env(), verbose = T)
}
\arguments{
\item{data}{A data.frame in which rows represent articles and columns represent data fields. See details for information about column names.}

\item{project}{The id of the project to add the articles to.}

\item{articleset}{The article set id of an existing set,}

\item{new_set}{The name of a new set to create a new articles. Note that passing the name of an existing set will create a new set with the same name. For uploading to an existing set, use the 'articleset' argument.}

\item{provenance}{If a new articleset is created (articleset is a name), an optional provenance string to store with the new set. If NULL, a default string mentioning the number of articles and creation date is used.}

\item{conn}{The connection object from \code{\link{amcat_connect}}.}

\item{verbose}{If true, report progress}
}
\value{
the articleset id
}
\description{
Upload a data.frame with article data to an articleset in AmCAT. Article data must be provided as a data.frame, 
in which every row represents an article, and every column represents a data field. All columns will be uploaded to AmCAT.

There are some \strong{special fields}, and it is recommended to use specific column names if you have this type of data.
Most importantly, the \strong{date} column is mandatory. Other common fields are \strong{text}, \strong{title}, \strong{byline}, \strong{url}, \strong{medium}, \strong{section} and \strong{author}.
See details (below) for more information.

All other columns in data will also be uploaded. By default, these columns will be stored as text, but specific field types (int, date, id, etc.) can be specified in the column name.
See details for the most imporant field types and how to use them.
}
\details{
The following field names have special meaning in AmCAT, and are recommended if you have this type of data.

\tabular{ll}{
  \strong{_date}    \tab Calendar date. Use a date type vector (Date, POSIXct or POSIXlt) or character type in format "%Y-%m-%d" or "%Y-%m-%d %H:%S:%M".   \cr
  \strong{text}     \tab The main body. Indexed for query searches. \cr
  \strong{title}    \tab The document title. Displayed above the text and indexed. \cr
  \strong{byline}   \tab A byline. Displayed below the title and indexed. \cr
  \strong{url}      \tab The URL, stored as a (clickable) keyword. \cr
  \strong{medium}   \tab The source of a document. \cr 
  \strong{section}  \tab A sub-category \cr
  \strong{author}   \tab The author of the document \cr
}

For alternative fields, it is possible (not required) to specify field types. To do so, add the type to the field name (column name in 'data'), separated by an underscore.
For example, to make the column "page" an integer column, use "page_int". The most important field types are:

\tabular{ll}{
  \strong{_int}     \tab Integers          \cr
  \strong{_date}    \tab Calendar date. Use a date type vector (Date, POSIXct or POSIXlt) or character type in format "%Y-%m-%d" or "%Y-%m-%d %H:%S:%M".   \cr
  \strong{_num}     \tab Numeric (double)  \cr
  \strong{_id}      \tab Keyword. For names and structured content such as zip codes \cr
  \strong{_url}     \tab Special keyword for URLs          \cr
  \strong{_text}    \tab Full text              \cr
  \strong{default}  \tab Without an explicit field type, field is stored as text \cr
}
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl', 'username')

set_id = create_articles(project = 1, articleset = 'example set', 
                         text = 'example', date = as.POSIXct('2010-01-01'), medium = 'example', headline = 'headline')
                         
get_articles(project = 1, articleset = set_id, 
             columns = c('headline','text','date','medium'))
}
}
