% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{create_articleset}
\alias{create_articleset}
\title{Create a new articleset on AmCAT}
\usage{
create_articleset(project, name, provenance = NULL,
  conn = conn_from_env())
}
\arguments{
\item{project}{The project id}

\item{name}{The name of the new set}

\item{provenance}{A provenance message. If NULL, a default message noting R and the time of creation is used.}

\item{conn}{The connection object from \code{\link{amcat_connect}}.}
}
\value{
The id of the new articleset
}
\description{
Create a new articleset on AmCAT
}
