% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{create_project}
\alias{create_project}
\title{Create a new project in AmCAT}
\usage{
create_project(name, description, owner = NULL, active = T,
  guest_role = c("metareader", "reader", "read/write", "admin", "none"),
  conn = conn_from_env())
}
\arguments{
\item{name}{The name of the new set}

\item{description}{A brief description}

\item{owner}{Specify the user that owns the project. Either provide an existing username (character) or user id (numeric). If NULL, 
the username of the current connection is used.}

\item{active}{If TRUE, set project as active (which makes a lot of sense for a new project)}

\item{guest_role}{Determine the priviledges of guests (i.e. users not assigned to the project). See details for options.}

\item{conn}{The connection object from \code{\link{amcat_connect}}. 

\tabular{ll}{
  \strong{metareader}  \tab Guest can see and query articlesets, but only sees meta data and snippets (not full texts) \cr
  \strong{reader}      \tab The above, plus read articles, and copy articlesets to their own project \cr
  \strong{read/write}  \tab The above, plus create and modify articlesets  \cr
  \strong{admine}      \tab The above, plus managing user privileges and deleting articlesets \cr
  \strong{none}        \tab Give no access. \cr
}}
}
\value{
The id of the new project
}
\description{
Create a new project in AmCAT
}
