% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{create_user}
\alias{create_user}
\title{Create a new user on AmCAT}
\usage{
create_user(project, name, provenance = NULL, conn = conn_from_env())
}
\arguments{
\item{provenance}{A provenance message. If NULL, a default message noting R and the time of creation is used.}

\item{conn}{The connection object from \code{\link{amcat_connect}}.}

\item{username}{The username on AmCAT. Max 30 characters. Letters, digits and [@.+-_] only.}

\item{password}{The}

\item{first_name}{The first name of the new user}

\item{last_name}{The last name of the new user}

\item{email}{The user's email adress}

\item{active}{If False, users are not considered 'active'. Unselect this instead of deleting accounts.}

\item{superuser}{If TRUE, make user a superuser}

\item{staff}{If TRUE, give user staff status}
}
\value{
The id of the new articleset
}
\description{
If the current user is a superuser, new users can be added.
}
