% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/django.r
\name{DecimalField}
\alias{DecimalField}
\title{Create a Django Form Field}
\usage{
DecimalField(label = NA, initial = 0, required = T)
}
\description{
Use to create field objects for R plugins in AmCAT. See djangoFormFields().
}
