% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/django.r
\name{djangoFormFields}
\alias{djangoFormFields}
\title{Create a json string for django form fields}
\usage{
djangoFormFields(...)
}
\arguments{
\item{...}{the field objects, created using the functions: BooleanField, IntegerField, etc. Note that the field objects have to be named!}
}
\value{
a json string
}
\description{
Create the form fields to be used for R plugins in AmCAT.
}
\examples{
djangoFormFields(field1=BooleanField(initial=F, required=F),
                 field2=IntegerField(initial=10, required=T))
}
