% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.r
\name{.create.cats}
\alias{.create.cats}
\title{Create categories from a hierarchy (parent-code) list
The result is a data frame with the same number of rows as the hierarchy,
where the first column consists only of the roots, the second column of the roots and their children
and so un until the last column consists of all nodes, so each row is a path to the root}
\usage{
.create.cats(hierarchy, label = "code", maxdepth = NULL)
}
\arguments{
\item{label}{column names to generate: label, label.p1, label.p2 etc}

\item{maxdepth}{number of columns to keep, not counting the code column.}
}
\description{
Create categories from a hierarchy (parent-code) list
The result is a data frame with the same number of rows as the hierarchy,
where the first column consists only of the roots, the second column of the roots and their children
and so un until the last column consists of all nodes, so each row is a path to the root
}
