% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.r
\name{.gethierarchy}
\alias{.gethierarchy}
\title{Memoised function to get the hierarchy as code / parent columns, both character "<ID> - <Label>"}
\usage{
.gethierarchy(conn, codebook_id)
}
\description{
Memoised function to get the hierarchy as code / parent columns, both character "<ID> - <Label>"
}
