% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{flush_elasticsearch}
\alias{flush_elasticsearch}
\title{Ask AmCAT to flush the elasticsearch}
\usage{
flush_elasticsearch(conn)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat_connect}}}
}
\description{
Ask AmCAT to flush the elasticsearch
}
