% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tokens.r
\name{flush_token}
\alias{flush_token}
\title{Delete saved AmCAT tokens for given host and username}
\usage{
flush_token(host, username)
}
\arguments{
\item{host}{the hostname, e.g. http://amcat.vu.nl or http://localhost:8000}

\item{username}{an existing username on the host server}
}
\description{
Delete saved AmCAT tokens for given host and username
}
