% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_aggregate}
\alias{get_aggregate}
\title{Conduct an aggregate query on amcat}
\usage{
get_aggregate(queries, sets, conn = conn_from_env(), labels = queries,
  axis1 = NULL, axis2 = NULL, ...)
}
\arguments{
\item{queries}{a vector of queries to run}

\item{sets}{one or more article set ids to query on}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{labels}{if given, labels corresponding to the queries}

\item{axis1}{The first grouping (break/group by) variable, e.g. year, month, week, day, or medium}

\item{axis2}{The second grouping (break/group by) variable, e.g. medium. Do not use a date interval here.}

\item{...}{additional arguments to pass to the AmCAT API.}
}
\value{
A data frame with hits per group
}
\description{
This function is similar to using the 'show table' function in AmCAT. It allows you to specify a
number of queries and get the number of hits per search term, per period, etc.
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl')
a = get_aggregate(sets=10271, axis1="year",
                  queries=c("*", "obama", "bush"), 
                  labels=c("Total", "Obama", "Bush"))
head(a)
}
}
