% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{get_articles}
\alias{get_articles}
\title{Get article data from AmCAT}
\usage{
get_articles(project, articleset = NULL, articles = NULL,
  conn = conn_from_env(), uuid = NULL, columns = c("id", "date", "title"),
  include_all = F, include_text = F, time = F, dateparts = F,
  page_size = 10000, ...)
}
\arguments{
\item{project}{the project of a set to retrieve metadata from}

\item{articleset}{the article set id to retrieve - provide either this or articleset}

\item{articles}{the article ids to retrieve - provide either this or articleset}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{uuid}{like the articles argument, but using the universally unique identifier (uuid)}

\item{columns}{the names of columns to retrieve, e.g. date, medium, text, headline. If set to NULL, all available columns are retrieved.}

\item{include_all}{shorthand for including all columns. (this is identical to passing NULL to columns)}

\item{include_text}{shorthand for including the "headline", "byline" and "text" columns. (this is identical to passing these names to columns)}

\item{time}{if true, parse the date as POSIXct datetime instead of Date}

\item{dateparts}{if true, add date parts (year, month, week)}

\item{page_size}{the number of articles per (downloaded) page}
}
\value{
A data.frame with article columns
}
\description{
Retrieve article data. Be default only returns basic meta data ("id","date","title"),
but any available article data can be retrieved, including text.
}
\details{
If "date" is included, it will be returned as either a Date (default) or as a POSIXct (if time is TRUE)
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl', 'username')
meta = get_articles(project = 1, articleset = 1)
head(meta) 

texts = get_articles(project = 1, articleset = 1, columns = c('text'))
texts$text[1]
}
}
