% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.r
\name{get_articlesets}
\alias{get_articlesets}
\title{View projects}
\usage{
get_articlesets(project, only_favourite = F, columns = c("id", "project",
  "name", "featured", "favourite", "provenance"), conn = conn_from_env())
}
\arguments{
\item{project}{The project id}

\item{only_favourite}{If TRUE, only "favourite" projects are shown}

\item{columns}{the names of columns to retrieve, e.g. date, medium, text, headline. If set to NULL, all available columns are retrieved}

\item{conn}{the connection object from \code{\link{amcat_connect}}}
}
\value{
A data.frame with articlesets
}
\description{
View projects
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl', 'username')
get_articlesets(project=1)
}
}
