% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.r
\name{get_corpus}
\alias{get_corpus}
\title{Get article metadata from AmCAT}
\usage{
get_corpus(project, articleset = NULL, articles = NULL,
  conn = conn_from_env(), uuid = NULL, text_columns = c("headline",
  "byline", "text"), meta_columns = NULL, time = F, dateparts = F,
  page_size = 10000, format = c("quanteda_corpus", "tcorpus"), ...)
}
\arguments{
\item{project}{the project of a set to retrieve metadata from}

\item{articleset}{the article set id to retrieve - provide either this or articleset}

\item{articles}{the article ids to retrieve - provide either this or articleset}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{uuid}{like the articles argument, but using the universally unique identifier (uuid)}

\item{text_columns}{The columns that contain the article text. If multiple columns are given (e.g., headline, text), they are pasted together (separated by a double linebreak)}

\item{meta_columns}{the names of columns to retrieve, e.g. date, medium, text, headline}

\item{time}{if true, parse the date as POSIXct datetime instead of Date}

\item{dateparts}{if true, add date parts (year, month, week)}

\item{page_size}{the number of articles per (downloaded) page}

\item{format}{determine whether article data is returned as a "data.frame", "quanteda_corpus" (requires quanteda package) or "tcorpus" (requires corpustools package).}

\item{...}{additional arguments are passed to quanteda::corpus or corpustools::create_tcorpus.}
}
\description{
Uses the \code{\link{get_objects}} function to retrieve article metadata, and applies some
additional postprocessing, e.g. to convert the data to Date objects.
}
