% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{get_objects}
\alias{get_objects}
\title{Get objects from the AmCAT API}
\usage{
get_objects(resource, conn = conn_from_env(), ...)
}
\arguments{
\item{resource}{the name of the resource, e.g. 'projects'. If it is of length>1, a path a/b/c/ will be created (e.g. c("projects",1,"articlesets"))}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{...}{Other options to pass to \code{\link{get_pages}}, e.g. page_size, format, and filters}
}
\value{
A dataframe of objects (rows) by properties (columns)
}
\description{
Get a table of objects from the AmCAT API, e.g. projects, sets etc.
}
