% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{get_pages}
\alias{get_pages}
\title{A wrapper for get_url to get data from multiple pages}
\usage{
get_pages(path, conn, param = list(), page_size = 500, start_page = 1,
  end_page = NA, verbose = T, ...)
}
\arguments{
\item{path}{the path of the url to retrieve (using the host from conn)}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{param}{a named vector of parameters, e.g. c(project=2, articleset=3)}

\item{page_size}{an integer, specifying the number of items per page}

\item{verbose}{if TRUE, report size of data to be retrieved and report progress}

\item{...}{additional parameters (param) in name-value pairs}
}
\value{
the raw result
}
\description{
A wrapper for get_url to get data from multiple pages
}
