% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.r
\name{get_projects}
\alias{get_projects}
\title{View projects}
\usage{
get_projects(only_visited = T, only_favourite = F, columns = c("id",
  "name", "active", "favourite", "weeks_ago", "guest_role"),
  conn = conn_from_env())
}
\arguments{
\item{only_visited}{If TRUE, only projects that the user visited will be shown}

\item{only_favourite}{If TRUE, only "favourite" projects are shown}

\item{columns}{the names of columns to retrieve, e.g. date, medium, text, headline. If set to NULL, all available columns are retrieved}

\item{conn}{the connection object from \code{\link{amcat_connect}}}
}
\value{
A data.frame with projects
}
\description{
View projects
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl', 'username')
get_projects()
}
}
