% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{get_url}
\alias{get_url}
\title{Retrieve a single URL from AmCAT with authentication and specified filters (GET or POST)}
\usage{
get_url(path, conn = conn_from_env(), filters = NULL, post = FALSE,
  post_options = list(), error_unless_200 = TRUE, null_on_404 = FALSE,
  binary = FALSE)
}
\arguments{
\item{path}{the path of the url to retrieve (using the host from conn)}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{filters}{a named vector of filters, e.g. c(project=2, articleset=3)}

\item{post}{use HTTP POST instead of GET}

\item{post_options}{a list with options for HTTP POST (if post is TRUE)}

\item{error_unless_200}{if TRUE, yield an error if HTTP request is not succesfull (status code 200: OK)}

\item{null_on_404}{if TRUE, return NULL if url not found (status code 404: NOT FOUND)}

\item{binary}{if TRUE, convert with \link{rawToChar}}
}
\value{
the raw result
}
\description{
Retrieve a single URL from AmCAT with authentication and specified filters (GET or POST)
}
