% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.r
\name{list.to.df}
\alias{list.to.df}
\title{Convert a list of rows (named list) into a data frame}
\usage{
list.to.df(x, stringsAsFactors = F)
}
\description{
Convert a list of rows (named list) into a data frame
}
