% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.r
\name{lucene_to_re}
\alias{lucene_to_re}
\title{Change a (simple) lucene query to a regular expression}
\usage{
lucene_to_re(queries, names = NULL)
}
\arguments{
\item{queries}{a character vector containing the queries}

\item{names}{an optional vector of the same length as queries containing names for the concepts}
}
\value{
a character vector of regular expressions, with names added if given
}
\description{
It is assumed that this will be used to filter single words/lemmata, so the query should only consist of a list of synonyms with optional wildcards.
E.g. a query like (many* words*) is fine, but many AND words or "many words" will give an error.
}

