% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.r
\name{make_pages_unique}
\alias{make_pages_unique}
\title{Make the given columns unique within a context for a list of data frames
 
It will iterate over the list of data frames and make the given columns globally unique
assuming they are unique within a context.
E.g. if you have a list of pages containing tokens within articles, and sentence is unique
within an article but not globally unique, this will make it globally unique.}
\usage{
make_pages_unique(result, context = "aid", columns = c("clause_id",
  "source_id", "coref", "sentence"))
}
\arguments{
\item{result}{list of token data frames}

\item{context}{name of the context column}

\item{columns}{names of the columns to make unique}
}
\value{
the original list with the columns made unique
}
\description{
Make the given columns unique within a context for a list of data frames
 
It will iterate over the list of data frames and make the given columns globally unique
assuming they are unique within a context.
E.g. if you have a list of pages containing tokens within articles, and sentence is unique
within an article but not globally unique, this will make it globally unique.
}
