% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_removed.r
\name{match_words}
\alias{match_words}
\title{Matches a regular expression pattern on multiple strings}
\usage{
match_words(pattern, words, ignore.case = T)
}
\arguments{
\item{pattern}{a regular expression pattern to use}

\item{words}{a vector of words, this should probably be a factor to achieve speedup}

\item{ignore.case}{passed to grepl, defaults to TRUE for this function}
}
\value{
a vector of booleans indicating which words matched the pattern
}
\description{
This is intended as a 'drop-in' replacement for grepl and should give identical results
This function first matches on the unique strings and then matches back to the original,
so it will give a speedup iff length(unique(words)) << length(words)
}
