% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{print.amcat_connection}
\alias{print.amcat_connection}
\title{S3 print method for amcat_connection (API connection) objects}
\usage{
\method{print}{amcat_connection}(x, ...)
}
\arguments{
\item{x}{an amcat_connection connection object, created with \link{amcat_connect}}

\item{...}{not used}
}
\description{
S3 print method for amcat_connection (API connection) objects
}
\examples{
\dontrun{
conn = amcat_connect('http://localhost:8000')
conn
}
}
