% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{print.amcatConnection}
\alias{print.amcatConnection}
\title{S3 print method for amcatConnection (API connection) objects}
\usage{
\method{print}{amcatConnection}(x, ...)
}
\arguments{
\item{x}{an amcatConnection connection object, created with \link{amcat_connect}}

\item{...}{not used}
}
\description{
S3 print method for amcatConnection (API connection) objects
}
\examples{
\dontrun{
conn = amcat_connect('http://localhost:8000')
conn
}
}
