% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.r
\name{process.codings}
\alias{process.codings}
\title{Process a 'long' list of raw codings into a 'wide' data frame with a column per coded variable}
\usage{
process.codings(codings, fields, codes)
}
\description{
Process a 'long' list of raw codings into a 'wide' data frame with a column per coded variable
}
