% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{publisher.column}
\alias{publisher.column}
\title{Return the column for publisher/medium depending om AmCAT version}
\usage{
publisher.column(conn)
}
\description{
Return the column for publisher/medium depending om AmCAT version
}
