% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_removed.r
\name{quanteda.corpus}
\alias{quanteda.corpus}
\title{Download articles from AmCAT into a Quanteda corpus object}
\usage{
quanteda.corpus(conn, project, articleset, textcolumns = c("headline",
  "text"), metacolumns = c("date", "medium"), headline.as.docvar = T, ...)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{project}{The project ID}

\item{articleset}{The Articleset ID}

\item{textcolumns}{Which columns contain text (default: headline, text)}

\item{metacolumns}{Which columns to include as docvars (deafult: date, medium)}

\item{headline.as.docvar}{Should headline be a docvar (as well as a text column)}

\item{...}{Other arguments to pass to get_articles, i.e. dateparts=T}
}
\value{
a quanteda corpus object
}
\description{
Download articles from AmCAT into a Quanteda corpus object
}
