% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{save_amcat_password}
\alias{save_amcat_password}
\title{Add or change an entry in the cached authentication file}
\usage{
save_amcat_password(host, username, password, passwordfile = "~/.amcatauth")
}
\arguments{
\item{host}{the host, e.g. "https://amcat.nl"}

\item{username}{the AmCAT username to use}

\item{password}{the password for the AmCAT user}

\item{passwordfile}{optionally, specify a different password file}
}
\description{
amcat-r uses the file '~/.amcatauth' to read credentials for connecting to servers
to prevent passwords from appearing in script files. This function will update
the .amcatauth file (if present) to add or change the password for the given host
}
\examples{
\dontrun{
host = 'https://amcat.nl'  ## existing and available host
username = 'XXX'           ## registered username  
password = 'XXX'           ## registered password

save_amcat_password(host = host, username = username, password = password)
}
}
