% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{title.column}
\alias{title.column}
\title{Return the column for title/headline depending om AmCAT version}
\usage{
title.column(conn)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}
}
\description{
Return the column for title/headline depending om AmCAT version
}
