% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{upload_articles}
\alias{upload_articles}
\title{Upload new articles to AmCAT}
\usage{
upload_articles(project, articleset, text, headline, date, medium,
  conn = conn_from_env(), provenance = NULL, ...)
}
\arguments{
\item{project}{the project to add the articles to}

\item{articleset}{the article set id of an existing set, or the name of a new set to create}

\item{text}{the text of the articles to upload}

\item{headline}{the headlines of the articles to upload}

\item{date}{the date of the articles to upload}

\item{medium}{the medium of the articles to upload.}

\item{conn}{the connection object from \code{\link{amcat_connect}}.}

\item{provenance}{if articleset is character, an optional provenance string to store with the new set}

\item{...}{and additional fields to upload, e.g. author, byline etc.}
}
\description{
Upload articles into a given project and article set, or into a new article set if the articleset argument is character
All arguments headline, medium etc. should be either of the same length as text, or of length 1
All factor arguments will be converted to character using as.character
For date, please provide either a string in ISO notatoin (i.e. "2010-12-31" or "2010-12-31T23:59:00")
or a variable that can be converted to string using format(), e.g. Date, POSIXct or POSIXlt. 
The articles will be uploaded in batches of 100.
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl')

set_id = upload_articles(project = 1, articleset = 'example set', 
                         headline = 'example', text = 'example',
                         date = as.POSIXct('2010-01-01'), medium = 'example')
                         
get_articles(project = 1, articleset = set_id, 
             columns = c('headline','text','date','medium'))
}
}
