% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.r
\name{amcat.aggregate}
\alias{amcat.aggregate}
\title{Conduct an aggregate query on amcat}
\usage{
amcat.aggregate(conn, queries, labels = queries, sets, axis1 = NULL,
  axis2 = NULL, ...)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{queries}{a vector of queries to run}

\item{labels}{if given, labels corresponding to the queries}

\item{sets}{one or more article set ids to query on}

\item{axis1}{The first grouping (break/group by) variable, e.g. year, month, week, day, or medium}

\item{axis2}{The second grouping (break/group by) variable, e.g. medium. Do not use a date interval here.}

\item{...}{additional arguments to pass to the AmCAT API.}
}
\value{
A data frame with hits per group
}
\description{
This function is similar to using the 'show table' function in AmCAT. It allows you to specify a
number of queries and get the number of hits per search term, per period, etc.
}

