% Generated by roxygen2 (4.0.1): do not edit by hand
\name{amcat.connect}
\alias{amcat.connect}
\title{Connect to the AmCAT API}
\usage{
amcat.connect(host, username = NULL, passwd = NULL, token = NULL,
  disable_ipv6 = TRUE)
}
\arguments{
\item{host}{the hostname, e.g. http://amcat.vu.nl or http://localhost:8000}

\item{username}{the username to login with, e.g. 'amcat'.}

\item{passwd}{the password to login with, e.g. 'amcat'}

\item{token}{an existing token to authenticate with. If given, username and password are not used and the token is not tested}

\item{disable_ipv6.}{If True, only use ipv4 resolving (faster if ipv6 causes timeout). Defaults to true, but this may change in the future.}
}
\value{
A list with authentication information that is used by the other functions in this package
}
\description{
Connect to the AmCAT API and requests a temporary (24h) authentication token that will be stored in the output
If username and password are not given, a file ~/.amcatauth will be read, which should be a csv file with
columns host (may be *), username, password. If the file cannot be read, username will be taken from $USER
and the user will be prompted for the password.
}

