% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{amcat.getarticlemeta}
\alias{amcat.getarticlemeta}
\title{Get article metadata from AmCAT}
\usage{
amcat.getarticlemeta(conn, set = NULL, article_ids = NULL,
  filters = list(), columns = c("id", "date", "medium", "length"),
  time = F, dateparts = F, medium_names = T, page_size = 10000, ...)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{set}{the article set id to retrieve}

\item{article_ids}{the article ids to retrieve}

\item{filters}{additional filters, e.g. list(medium=1)}

\item{columns}{the names of columns to retrieve}

\item{time}{if true, parse the date as POSIXct datetime instead of Date}

\item{dateparts}{if true, add date parts (year, month, week)}

\item{medium_names}{if true, retrieve medium names and turn medium column into a factor}

\item{...}{additional arguments are passed to \code{\link{amcat.getpages}}. Useful arguments include page_size, page (starting page) and maxpage (end page)}
}
\value{
A dataframe containing the articles and the selected columns
}
\description{
Uses the \code{\link{amcat.getobjects}} function to retrieve article metadata, and applies some
additional postprocessing, e.g. to convert the data to Date objects.
}

