% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{amcat.getarticlemeta}
\alias{amcat.getarticlemeta}
\title{Get article metadata from AmCAT}
\usage{
amcat.getarticlemeta(conn, project, articleset = NULL, articles = NULL,
  uuid = NULL, columns = c("date", "medium"), time = F, dateparts = F,
  page_size = 10000)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{project}{the project of a set to retrieve metadata from}

\item{articleset}{the article set id to retrieve}

\item{columns}{the names of columns to retrieve, e.g. date, medium, text, headline}

\item{time}{if true, parse the date as POSIXct datetime instead of Date}

\item{dateparts}{if true, add date parts (year, month, week)}

\item{medium_names}{if true, retrieve medium names and turn medium column into a factor}
}
\value{
A dataframe containing the articles and the selected columns
}
\description{
Uses the \code{\link{amcat.getobjects}} function to retrieve article metadata, and applies some
additional postprocessing, e.g. to convert the data to Date objects.
}
