% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corpus.r
\name{amcat.gettokens}
\alias{amcat.gettokens}
\title{Get Tokens from AmCAT}
\usage{
amcat.gettokens(conn, project = NULL, articleset = NULL,
  module = "corenlp_lemmatize", filters = NULL, page_size = 1, page = 1,
  npages = NULL, sentence = NULL, only_cached = F)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{project}{id of the project containing the tokens}

\item{articleset}{id of the articleset to get features from. If not specified, specify sentence for 'ad hoc' parsing}

\item{module}{the NLP preprocessing module to get the tokens from}

\item{filters}{Additional filters, ie c(pos1="V", pos1="A") to select only verbs and adjectives}

\item{page_size}{the number of features (articles?) to include per call}

\item{sentence}{a sentence (string) to be parsed if articleset id is not given}

\item{only_cached}{if true, only get tokens that have already been preprocessed (recommended for large corpora!)}
}
\value{
A data frame of tokens
}
\description{
Get Tokens (pos, lemma etc) from AmCAT
}

