% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{amcat_connect}
\alias{amcat_connect}
\title{Connect to the AmCAT API}
\usage{
amcat_connect(host = NULL, username = NULL, token = NULL,
  disable_ipv6 = TRUE, ssl.verifypeer = TRUE, conn = NULL)
}
\arguments{
\item{host}{the hostname, e.g. https://amcat.nl or http://localhost:8000}

\item{username}{an existing username on the host server}

\item{token}{an existing token to authenticate with. If given (and valid), no password has to be entered}

\item{disable_ipv6}{If True, only use ipv4 resolving (faster if ipv6 causes timeout). Defaults to true, but this may change in the future.}

\item{ssl.verifypeer}{If True, verifies the authenticity of the peer's certificate}

\item{conn}{an existing AmCAT connection object. If provided, the connection will be opened (if not yet open) and refreshed (i.e. new token). The other arguments are ignored.}
}
\value{
A S3 class (amcatConnection) with authentication information that is used by the other functions in this package
}
\description{
Connect to the AmCAT API and request or refresh a temporary (48h) authentication token.
}
\details{
The token can be used in two ways. Firstly, after running amcat_connect, the token will automatically be used
until the R session is ended. This is recommended for occasional use. Alternatively, amcat_connect() returns
a connection object (amcatConnection) that can be passed to the conn argument in functions that require a connection. 
The amcatConnection object can be saved for use across sessions, so that the password does not have to be given to connect.
It is also possible to pass the amcatConnection object to amcat_connect, which will refresh the token and set it up for
default use within a session.
}
\examples{
\dontrun{
## create connection
amcat_connect('https://amcat.nl', 'username')

## optionally, get the amcatConnect object, e.g., to save it for later use
conn = amcat_connect('https://amcat.nl', 'username')
saveRDS(conn, '~/.amcat_connection.rds')

## refresh connection 
conn = readRDS('~/.amcat_connection.rds')
conn = amcat_connect(conn = conn)
}
}
