% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{conn_from_env}
\alias{conn_from_env}
\title{Get the current AmCAT connection}
\usage{
conn_from_env()
}
\value{
an amcatConnection connection object, created with \link{amcat_connect}
}
\description{
Get the current AmCAT connection
}
