% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{get_articles}
\alias{get_articles}
\title{Get article metadata from AmCAT}
\usage{
get_articles(conn, project, articleset = NULL, articles = NULL,
  uuid = NULL, columns = c("date", "medium"), time = F, dateparts = F,
  page_size = 10000)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{project}{the project of a set to retrieve metadata from}

\item{articleset}{the article set id to retrieve - provide either this or articleset}

\item{articles}{the article ids to retrieve - provide either this or articleset}

\item{uuid}{like the articles argument, but using the universally unique identifier (uuid)}

\item{columns}{the names of columns to retrieve, e.g. date, medium, text, headline}

\item{time}{if true, parse the date as POSIXct datetime instead of Date}

\item{dateparts}{if true, add date parts (year, month, week)}

\item{page_size}{the number of articles per (downloaded) page}
}
\value{
A dataframe containing the articles and the selected columns
}
\description{
Uses the \code{\link{get_objects}} function to retrieve article metadata, and applies some
additional postprocessing, e.g. to convert the data to Date objects.
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl')
meta = get_articles(conn, project = 1, articleset = 1)
head(meta) 

texts = get_articles(conn, project = 1, articleset = 1, columns = c('text'))
texts$text[1]
}
}
