% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.r
\name{get_articles}
\alias{get_articles}
\title{Get article meta data from AmCAT}
\usage{
get_articles(project, articleset = NULL, articles = NULL,
  uuid = NULL, columns = c("date", "medium", "title"), text = F,
  time = F, dateparts = F, page_size = 10000, ignore_missing = T,
  conn = conn_from_env(), verbose = T, ...)
}
\arguments{
\item{project}{the project of a set to retrieve metadata from}

\item{articleset}{the article set id to retrieve - provide either this or articleset}

\item{articles}{the article ids to retrieve - provide either this or articleset (only works on latest AmCAT versions)}

\item{uuid}{like the articles argument, but using the universally unique identifier (uuid)}

\item{columns}{the names of columns to retrieve, e.g. date, medium, text, headline. If set to NULL, all available columns are retrieved.}

\item{text}{if TRUE, include the text field (text)}

\item{time}{if true, parse the date as POSIXct datetime instead of Date}

\item{dateparts}{if true, add date parts (year, month, week)}

\item{page_size}{the number of articles per (downloaded) page}

\item{ignore_missing}{If FALSE, requested columns (in the columns argument) that are not in the data will be added with NA values.}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{verbose}{If true, report progress}
}
\value{
A data.frame with article columns
}
\description{
Retrieve article data. Be default only returns basic meta data ("id","date","title"),
but any available article data can be retrieved.
}
\details{
For working with full texts, it is recommended to use the \link{amcat_corpus} function instead

If "date" is included, it will be returned as either a Date (default) or as a POSIXct (if time is TRUE)
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl', 'username')
meta = get_articles(project = 1, articleset = 1)
head(meta) 

texts = get_articles(project = 1, articleset = 1, columns = c('text'))
texts$text[1]
}
}
