% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{get_hits}
\alias{get_hits}
\title{Conduct a query on amcat}
\usage{
get_hits(queries, sets, conn = conn_from_env(), labels = queries,
  minimal = T, warn.no.result = T, ...)
}
\arguments{
\item{queries}{a vector of queries to run}

\item{sets}{one or more article set ids to query on}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{labels}{if given, labels corresponding to the queries. Alternatively, if a query starts with a label and a hashtag (e.g., label# "term1 term2"), this label is used.}

\item{minimal}{if TRUE, returns only hits per document. if set to FALSE, includes more meta data}

\item{warn.no.result}{if TRUE, yield a warning in case no results were found}

\item{...}{additional arguments to pass to the AmCAT API, e.g. extra filters}
}
\value{
A data frame with hits per article
}
\description{
This function is similar to using the 'show article list' function in AmCAT. It allows you to specify a
number of queries and get document metadata and number of hits per document
}
\examples{
\dontrun{
conn = amcat_connect('https://amcat.nl')
h = get_hits(queries=c("tyrant OR brute", "saddam"), labels=c("tyrant", "saddam"), sets=10271)
head(h)
table(h$query)
}
}
