% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{get_pages}
\alias{get_pages}
\title{Get and rbind pages from the AmCAT API}
\usage{
get_pages(path, conn = conn_from_env(), format = NULL, page = 1,
  page_size = 1000, filters = NULL, post = FALSE, post_options = list(),
  max_page = NULL, rbind_results = T)
}
\arguments{
\item{path}{the path of the url to retrieve (using the host from conn)}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{format}{a character string giving the output format. Can be 'rda', 'csv' or 'json'. 'rda' is recommended. If NULL (default), 'rda' is used.}

\item{page}{the page number to start retrieving}

\item{page_size}{the number of rows per page}

\item{filters}{a named vector of filters, e.g. c(project=2, articleset=3)}

\item{post}{use HTTP POST instead of GET}

\item{post_options}{a list with options for the HTTP POST (if post is TRUE)}

\item{max_page}{the page number to stop retrieving at, if given}

\item{rbind_results}{if TRUE, return results of multiple pages as a single data.frame. if FALSE, a list with data.frames is returned}
}
\value{
dataframe
}
\description{
Get and rbind pages from the AmCAT API
}
