% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{get_url}
\alias{get_url}
\title{Retrieve a single URL from AmCAT with authentication and specified filters (GET or POST)}
\usage{
get_url(path, conn, param = list(), post = FALSE, post_options = list(),
  read = T, ...)
}
\arguments{
\item{path}{the path of the url to retrieve (using the host from conn)}

\item{conn}{the connection object from \code{\link{amcat_connect}}}

\item{param}{a named vector of parameters, e.g. c(project=2, articleset=3)}

\item{post}{use HTTP POST instead of GET}

\item{post_options}{a list with options for HTTP POST (if post is TRUE)}

\item{read}{If TRUE, read the content from the response. Otherwise, return the response object}

\item{...}{additional parameters (param) in name-value pairs}
}
\value{
the raw result
}
\description{
Retrieve a single URL from AmCAT with authentication and specified filters (GET or POST)
}
