% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.r
\name{scroll}
\alias{scroll}
\title{Scroll an amcat API page with 'next' link}
\usage{
scroll(conn, path, page_size = 100, ...)
}
\arguments{
\item{conn}{the connection object from \code{\link{amcat.connect}}}

\item{path}{the path to scroll}

\item{page_size}{the amount of pages per request}

\item{...}{additional query arguments}
}
\value{
a data frame of all returned rows
}
\description{
Scroll an amcat API page with 'next' link
}
